Revert "troubleshoot route issues, cannot login"
This reverts commit 2ecf59483ebb078a266dce11576dc0513d4e35c9.

+++ public/js/delete-post.js
@@ -1,8 +1,9 @@
 // deletion handler: send DELETE request to API if user clicks delete
 const deletePostHandler = async (event) => {
-  const id = event.target.getAttribute('data-id');
-
   try {
+    if (event.target.hasAttribute('data-id')) {
+      const id = event.target.getAttribute('data-id');
+
       const response = await fetch(`/api/dashboard/${id}`, {
         method: 'DELETE',
       });
@@ -15,6 +16,7 @@ const deletePostHandler = async (event) => {
         alert('Failed to delete blog post');
         console.error(response.statusText);
       }
+    }
   } catch (err) {
       alert('An error occurred. Please try again.');
       console.error(err);
@@ -23,6 +25,4 @@ const deletePostHandler = async (event) => {
 };
 
 // update bloglist on dashboard upon submit
-document
-  .querySelector('.edit-post-form')
-  .addEventListener('submit', deletePostHandler);
+document.querySelector('.user-blog-list').addEventListener('submit', deletePostHandler);
